Navigation enabled
Now uses navigator and stack screens. Has necessary node modules.
