+++ package.json
@@ -9,6 +9,9 @@
     "web": "expo start --web"
   },
   "dependencies": {
+    "@react-navigation/native": "^6.1.6",
+    "@react-navigation/native-stack": "^6.9.12",
+    "@react-navigation/stack": "^6.3.16",
     "expo": "~48.0.10",
     "expo-camera": "~13.2.1",
     "expo-cli": "^6.3.2",
@@ -17,8 +20,11 @@
     "react": "18.2.0",
     "react-dom": "^18.2.0",
     "react-native": "0.71.6",
-    "react-native-vision-camera": "^2.15.4",
-    "react-native-web": "~0.18.11"
+    "react-native-gesture-handler": "^2.9.0",
+    "react-native-safe-area-context": "4.5.0",
+    "react-native-screens": "~3.20.0",
+    "react-native-web": "~0.18.11",
+    "react-native-reanimated": "~2.14.4"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
